recipes/models.py
from django.db import models
from users.models import User


class Ingredient(models.Model):
    name = models.CharField(max_length=255, unique=True)

    class Meta:
        ordering = ['name']

    def __str__(self):
        return self.name


class Recipe(models.Model):
    name = models.CharField(max_length=255, unique=True)
    recipe_text = models.TextField()
    ingredients = models.ManyToManyField(Ingredient)
    user = models.ForeignKey(User, on_delete=models.CASCADE)

    class Meta:
        ordering = ['name']

    def __str__(self):
        return self.name


class Rate(models.Model):
    rate = models.PositiveSmallIntegerField(unique=True)

    class Meta:
        ordering = ['rate']


class Rating(models.Model):
    user = models.ForeignKey(User, on_delete=models.CASCADE)
    recipe = models.ForeignKey(Recipe, on_delete=models.CASCADE)
    rate = models.ForeignKey(Rate, on_delete=models.CASCADE)

    class Meta:
        ordering = ['user', 'recipe', 'rate']
        unique_together = ('user', 'recipe')
